// backup/src/types.tsx
export type TipeGender = 'L' | 'P';
export type TipeKelas = '1' | '2' | '3';
export type TipeJurusan = 'KITAB' | 'TAHFIDZ';
export type StatusSantri = 'AKTIF' | 'LULUS' | 'KELUAR' | 'ALUMNI';
// ...
export interface ISantri {
  nis: string; // PK
  wali_id: string | null; // FK to profiles
  nama: string;
  nik: string | null;
  tempat_lahir: string | null;
  tanggal_lahir: string; // Date string YYYY-MM-DD
  jenis_kelamin: TipeGender;
  kelas: TipeKelas;
  jurusan: TipeJurusan;
  status_santri: StatusSantri;
  foto_url: string | null;
  ayah: string | null;
  ibu: string | null;
  no_kontak_wali: string | null;
  alamat_lengkap: string | null;
  anak_ke: number | null;
  hafalan_kitab : string | null;
  total_hafalan : number | null; // in Juz
  pembimbing: string | null;
  created_at: string;
  latitude?: number | null;
  longitude?: number | null;
  kecamatan_id?: string | null;
}
// ...
export const NAMA_KITAB_LIST = [
    "Jurumiah", 
    "Imrity", 
    "Nadzmul Maqshud", 
    "Alfiyah", 
    "Uqudul Juman", 
    "Sulam Munawraq"
] as const;

export interface IHafalanKitab {
    id: number;
    created_at: string;
    santri_nis: string;
    santri?: ISantri;
    tanggal: string;
    nama_kitab: typeof NAMA_KITAB_LIST[number];
    bab_materi: string;
    bait_awal?: number;
    bait_akhir?: number;
    halaman_awal?: number;
    halaman_akhir?: number;
    predikat: 'MUMTAZ' | 'JAYYID' | 'KURANG';
    status: 'LULUS' | 'MENGULANG';
    catatan: string;
    dicatat_oleh_id: string;
}